+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/sharding/jdbc/unsupported/UnsupportedOperationStatementTest.java
@@ -60,11 +60,6 @@ public final class UnsupportedOperationStatementTest extends AbstractShardingDat
     }
     
     @Test(expected = SQLFeatureNotSupportedException.class)
-    public void assertGetGeneratedKeys() throws SQLException {
-        actual.getGeneratedKeys();
-    }
-    
-    @Test(expected = SQLFeatureNotSupportedException.class)
     public void assertAddBatch() throws SQLException {
         actual.addBatch("");
     }